Removed all unnecessary images and added the right videos for the project
