+++ index.html
@@ -6,6 +6,7 @@
         <meta name="description" content="Personal Portfolio Website of Felix Baah, a Software Developer based in London">
         <meta name="author" content="Felix Baah">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <link rel="icon" href="/src/assets/svg/portfolio-site-icon.svg" type="image/svg+xml" sizes="32x32">
 
         <!----------------------browser compatibility ---------------------->
         <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -25,7 +26,7 @@
         <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
         <link rel="preconnect" href="https://fonts.googleapis.com">
         <link href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&family=Raleway:ital,wght@0,100..900;1,100..900&family=Urbanist:ital,wght@0,100..900;1,100..900&display=swap" rel="stylesheet">
-        
+
         <!---------------------- fontAwesome icons ---------------------->
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.css"/>
 
@@ -52,10 +53,7 @@
         <meta name="robots" content="index, follow" />
 
         <!-- Add an icon for Apple devices -->
-        <link rel="apple-touch-icon" sizes="180x180" href="./assets/images/portfolio-site-icon.jpg">
-
-        <!---------------------- image icons on desktop ---------------------->
-        <link rel="icon" href="./asessts/images/portfolio-site-icon.jpg" type="image" sizes="32x32">
+        <link rel="apple-touch-icon" sizes="180x180" href="/src/assets/images/portfolio-site-icon.jpg">
 
         <!---------------------- title of page ---------------------->
         <title>Felix Baah | Personal Portfolio Website</title>
